api/utils/pagination_helpers.py
from ..pagination import StandardResultsSetPagination
from rest_framework.response import Response

def paginate_queryset(queryset, request, serializer_class):
    paginator = StandardResultsSetPagination()
    page = paginator.paginate_queryset(queryset, request)
    if page is not None:
        serializer = serializer_class(page, many=True)
        return paginator.get_paginated_response(serializer.data)
    serializer = serializer_class(queryset, many=True)
    return Response(serializer.data)
